Code cleanups - Moved color profile definitions to Colorprofiles.h, ditched cpp - Simplified access to profiles in header files
