+++ code/core/arm9/source/GbaColorCorrection/ColorLut.cpp
@@ -1,5 +1,4 @@
 #include "common.h"
-#include "ColorProfiles.h"
 #include "ColorLut.h"
 #include "GammaLut.h"
